+++ package.json
@@ -17,6 +17,7 @@
     "next": "latest",
     "react": "^18.2.0",
     "react-bootstrap": "^2.4.0",
+    "react-copy-to-clipboard": "^5.1.0",
     "react-dom": "^18.2.0",
     "react-modal": "^3.15.1",
     "react-query": "^3.39.1",
