+++ src/components/Sharing/SharingButtons.js
@@ -1,5 +1,3 @@
-import React from 'react'
-
 /* link sharing */
 import {
   EmailShareButton,
