+++ styles/index.css
@@ -252,6 +252,7 @@ a:hover,
 .sharing-icon:hover {
   transition: all 0.2s ease-in-out;
   filter: brightness(70%);
+  cursor: pointer;
 }
 
 /* search bar */
@@ -331,7 +332,7 @@ a:hover,
 .footer {
   display: flex;
   justify-content: space-between;
-  padding: 1rem 0.5rem;
+  padding: 1rem 2rem;
 }
 
 /* responsive */
